extract test into its own file
