+++ test/integration/broker.test.js
@@ -1,9 +1,8 @@
-const WebSocket = require('ws')
 const { startTracker } = require('streamr-network')
 const fetch = require('node-fetch')
 const { wait, waitForCondition } = require('streamr-test-utils')
 
-const { startBroker, createClient, getWsUrl } = require('../utils')
+const { startBroker, createClient } = require('../utils')
 
 const httpPort1 = 12341
 const httpPort2 = 12342
@@ -18,96 +17,6 @@ const trackerPort = 12370
 const broker1Key = '0x241b3f241b110ff7b3e6d52e74fea922006a83e33ff938e6e3cba8a460c02513'
 const broker2Key = '0x3816c1d1a81588cecf9ac271a4758ed08f208902c2dcda82ba1a2f458ac23a15'
 const broker3Key = '0xe8af31f5c61b64f44adcdab8c5c78a7bc0beea9dbf43af63f80544a1b84ec149'
-
-describe('websocket server', () => {
-    let ws
-    let broker
-
-    afterEach(async () => {
-        if (ws) {
-            ws.terminate()
-        }
-        await broker.close()
-    })
-
-    it('receives unencrypted connections', async (done) => {
-        broker = await startBroker({
-            name: 'broker1',
-            privateKey: broker1Key,
-            networkPort: networkPort1,
-            trackerPort,
-            httpPort: httpPort1,
-            wsPort: wsPort1
-        })
-        ws = new WebSocket(getWsUrl(wsPort1))
-        ws.on('open', async () => {
-            done()
-        })
-        ws.on('error', (err) => done(err))
-    })
-
-    it('receives encrypted connections', async (done) => {
-        broker = await startBroker({
-            name: 'broker1',
-            privateKey: broker1Key,
-            networkPort: networkPort1,
-            trackerPort,
-            httpPort: httpPort1,
-            wsPort: wsPort1,
-            privateKeyFileName: 'test/fixtures/key.pem',
-            certFileName: 'test/fixtures/cert.pem'
-        })
-        ws = new WebSocket(getWsUrl(wsPort1, true), {
-            rejectUnauthorized: false // needed to accept self-signed certificate
-        })
-        ws.on('open', async () => {
-            done()
-        })
-        ws.on('error', (err) => done(err))
-    })
-
-    describe('rejections', () => {
-        const testRejection = async (connectionUrl) => {
-            broker = await startBroker({
-                name: 'broker1',
-                privateKey: broker1Key,
-                networkPort: networkPort1,
-                trackerPort,
-                httpPort: httpPort1,
-                wsPort: wsPort1
-            })
-            ws = new WebSocket(connectionUrl)
-            let gotError = false
-            let closed = false
-            ws.on('open', () => {
-                throw new Error('Websocket should not have opened!')
-            })
-            ws.on('error', (err) => {
-                if (err.message.includes('400')) {
-                    gotError = true
-                } else {
-                    throw new Error(`Got unexpected error message: ${err.message}`)
-                }
-            })
-            ws.on('close', () => {
-                closed = true
-            })
-            await waitForCondition(() => gotError && closed)
-        }
-
-        it('rejects connections without preferred versions given as query parameters', async () => {
-            await testRejection(`ws://127.0.0.1:${wsPort1}/api/v1/ws`)
-        })
-
-        it('rejects connections with unsupported ControlLayer version', async () => {
-            await testRejection(getWsUrl(wsPort1, false, 666, 31))
-        })
-
-        it('rejects connections with unsupported MessageLayer version', async () => {
-            await testRejection(getWsUrl(wsPort1, false, 1, 666))
-        })
-    })
-})
 
 describe('broker: end-to-end', () => {
     let tracker
