+++ test/utils.js
@@ -52,16 +52,16 @@ function formConfig({
             hostname: '127.0.0.1',
             port: networkPort,
             advertisedWsUrl: null,
+            isStorageNode: true,
             trackers: [
                 `ws://127.0.0.1:${trackerPort}`
             ],
-            isStorageNode: true
-        },
-        location: {
-            latitude: 60.19,
-            longitude: 24.95,
-            country: 'Finland',
-            city: 'Helsinki'
+            location: {
+                latitude: 60.19,
+                longitude: 24.95,
+                country: 'Finland',
+                city: 'Helsinki'
+            }
         },
         cassandra: enableCassandra ? {
             hosts: ['localhost'],
